pyproject.toml
[tool.poetry]
name = "seldom-atx"
version = "0.1.0"
description = "seldom and openatx complete mobile automation test."
authors = ["bugmaster <fnngj@126.com>"]
license = "Apache-2.0"
readme = "README.md"

homepage = "https://github.com/SeldomQA/seldom-atx"
repository = "https://github.com/SeldomQA/seldom-atx"
documentation = "https://github.com/SeldomQA/seldom-atx"

keywords = ["sedom", "openatx", "app"]

classifiers = [
    'Intended Audience :: Developers',
    'Operating System :: MacOS :: MacOS X',
    'Operating System :: Microsoft :: Windows',
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    'Topic :: Internet :: WWW/HTTP :: Dynamic Content',
    'Topic :: Software Development :: Testing',
]

[tool.poetry.dependencies]
python = "^3.7"
XTestRunner = "^1.7.0"
parameterized = "~0.8.1"
loguru = "~0.6.0"
openpyxl = "^3.0.3"
pyyaml = "^6.0"
requests = "^2.22.0"
jsonschema = "^4.10.0"
jmespath = "^0.10.0"
pymysql = "^1.0.0"
genson = "~1.2.2"
click = "^8.1.3"
python-dateutil = "~2.8.2"
solox = "2.7.2"
facebook-wda = "1.4.6"
uiautomator2= "2.16.23"
matplotlib= "3.7.1"
gevent = "23.7.0"
pandas = "2.0.3"
xlsxwriter = "3.1.2"
tidevice = "0.11.0"

[tool.poetry.scripts]
seldom = "seldom.cli:main"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple/"
